+++ src/search_schemes/generator/zeroOnesZero.h
@@ -4,7 +4,7 @@
 
 namespace search_schemes::generator {
 
-auto zeroOnesZero_trivial(int minK, int K) -> Scheme;
-auto zeroOnesZero_opt(int minK, int K) -> Scheme;
+auto zeroOnesZero_trivial(size_t minK, size_t K) -> Scheme;
+auto zeroOnesZero_opt(size_t minK, size_t K) -> Scheme;
 
 }
